+++ locals.tf
@@ -2,6 +2,11 @@ locals {
   name_prefix  = var.name_prefix != "" ? replace(var.name_prefix, "/[a-z0-9]$/", "$0-") : ""
   default_name = lower("${local.name_prefix}${var.stack}-${var.client_name}-${var.location_short}-${var.environment}")
 
+  default_sku       = "GP_Gen5_2"
+  default_collation = "SQL_Latin1_General_CP1_CI_AS"
+
+  storage_account_type = "GRS"
+
   server_name = coalesce(var.server_custom_name, "${local.default_name}-sql")
 
   elastic_pool_name = coalesce(var.elastic_pool_custom_name, "${local.default_name}-pool")
@@ -17,8 +22,6 @@ locals {
     family   = contains(local.vcore_tiers, var.sku.tier) ? local.elastic_pool_vcore_family : null
   } : null
 
-  databases_users = var.create_databases_users ? { for db in var.elasticpool_databases : db => format("%s_user", replace(db, "-", "_")) } : {}
-
   allowed_subnets = [
     for id in var.allowed_subnets_ids : {
       name      = split("/", id)[10]
@@ -26,16 +29,34 @@ locals {
     }
   ]
 
-  default_database_single = {
-    collation      = "SQL_Latin1_General_CP1_CI_AS"
-    read_scale     = false
-    sku_name       = "Basic"
-    zone_redundant = false
-    threat_detection_policy = {
-      state = "Disabled"
+  databases_users = var.create_databases_users ? [
+    for db in var.databases_configuration : {
+      username = format("%s_user", replace(db.name, "-", "_"))
+      database = db.name
+      roles    = ["db_owner"]
     }
-    retention_days = 7
+  ] : []
+
+  standard_allowed_create_mode = {
+    "a" = "Default"
+    "b" = "Copy"
+    "c" = "Secondary"
+    "d" = "PointInTimeRestore"
+    "e" = "Restore"
+    "f" = "Recovery"
+    "g" = "RestoreExternalBackup"
+    "h" = "RestoreExternalBackup"
+    "i" = "RestoreLongTermRetentionBackup"
+    "j" = "OnlineSecondary"
   }
 
-  single_databases_configuration = [for db in var.single_databases_configuration : merge(local.default_database_single, { for k, v in db : k => v if v != null }) if var.enable_elasticpool == false]
+  datawarehouse_allowed_create_mode = {
+    "a" = "Default"
+    "b" = "PointInTimeRestore"
+    "c" = "Restore"
+    "d" = "Recovery"
+    "e" = "RestoreExternalBackup"
+    "f" = "RestoreExternalBackup"
+    "g" = "OnlineSecondary"
+  }
 }