+++ outputs.tf
@@ -11,18 +11,18 @@ output "sql_administrator_password" {
 }
 
 output "sql_server" {
-  description = "SQL Server"
-  value       = azurerm_sql_server.server
+  description = "SQL Server FQDN"
+  value       = azurerm_sql_server.server.fully_qualified_domain_name
 }
 
-output "sql_elastic_pool" {
-  description = "SQL Elastic Pool"
-  value       = azurerm_mssql_elasticpool.elastic_pool
+output "sql_server_id" {
+  description = "SQL Server ID"
+  value       = azurerm_sql_server.server.id
 }
 
 output "sql_databases" {
   description = "SQL Databases"
-  value       = azurerm_sql_database.db
+  value       = azurerm_mssql_database.db
 }
 
 output "sql_elastic_pool_id" {
@@ -32,23 +32,13 @@ output "sql_elastic_pool_id" {
 
 output "sql_databases_id" {
   description = "Map of the SQL Databases IDs"
-  value       = { for db in azurerm_sql_database.db : db.name => db.id }
-}
-
-output "sql_databases_creation_date" {
-  description = "Map of the SQL Databases creation dates"
-  value       = { for db in azurerm_sql_database.db : db.name => db.creation_date }
-}
-
-output "sql_databases_default_secondary_location" {
-  description = "Map of the SQL Databases default secondary location"
-  value       = { for db in azurerm_sql_database.db : db.name => db.default_secondary_location }
+  value       = { for db in azurerm_mssql_database.db : db.name => db.id }
 }
 
 output "default_administrator_databases_connection_strings" {
   description = "Map of the SQL Databases with administrator credentials connection strings"
   value = {
-    for db in azurerm_sql_database.db : db.name => formatlist(
+    for db in azurerm_mssql_database.db : db.name => formatlist(
       "Server=tcp:%s;Database=%s;User ID=%s;Password=%s;Encrypt=true;",
       azurerm_sql_server.server.fully_qualified_domain_name,
       db.name,
@@ -59,25 +49,44 @@ output "default_administrator_databases_connection_strings" {
   sensitive = true
 }
 
-output "databases_users" {
-  description = "Map of the SQL Databases dedicated usernames"
-  value       = local.databases_users
-  sensitive   = true
+output "default_databases_users" {
+  description = "Map of the SQL Databases dedicated users"
+  value = {
+    for db_user in local.databases_users :
+    "${db_user.username}-${db_user.database}" => db_user.username
+  }
 }
 
-output "databases_users_passwords" {
+output "default_databases_users_passwords" {
   description = "Map of the SQL Databases dedicated passwords"
-  value       = var.create_databases_users ? { for db in azurerm_sql_database.db : db.name => random_password.db_passwords[db.name].result } : {}
-  sensitive   = true
+  value = {
+    for db_user in local.databases_users :
+    "${db_user.username}-${db_user.database}" => random_password.db_passwords["${db_user.username}-${db_user.database}"].result
+  }
+  sensitive = true
 }
 
-output "custom_users_passwords" {
-  description = "Map of the custom users passwords"
-  value       = length(var.custom_users) == 0 ? {} : { for user in var.custom_users : format("%s-%s", user.name, user.database) => random_password.custom_users[format("%s-%s", user.name, user.database)].result }
-  sensitive   = true
+output "custom_databases_users" {
+  description = "Map of the custom SQL Databases users"
+  value = {
+    for custom_user in var.custom_users :
+    join("-", [custom_user.name, custom_user.database]) => module.custom_users[join("-", [custom_user.name, custom_user.database])].custom_user_name
+  }
 }
 
-output "databases_single_ids" {
-  description = "MSSQL Database single IDs map"
-  value       = { for db in azurerm_mssql_database.single_database : db.name => db.id }
+output "custom_databases_users_passwords" {
+  description = "Map of the custom SQL Databases users passwords"
+  value = {
+    for custom_user in var.custom_users :
+    join("-", [custom_user.name, custom_user.database]) => module.custom_users[join("-", [custom_user.name, custom_user.database])].custom_user_password
+  }
+  sensitive = true
+}
+
+output "custom_databases_users_roles" {
+  description = "Map of the custom SQL Databases users roles"
+  value = {
+    for custom_user in var.custom_users :
+    join("-", [custom_user.name, custom_user.database]) => module.custom_users[join("-", [custom_user.name, custom_user.database])].custom_user_roles
+  }
 }