+++ r-sql.tf
@@ -23,8 +23,9 @@ resource "azurerm_sql_firewall_rule" "firewall_rule" {
 }
 
 resource "azurerm_mssql_elasticpool" "elastic_pool" {
-  count = var.enable_elasticpool ? 1 : 0
-  name  = local.elastic_pool_name
+  count = var.elasticpool_enabled ? 1 : 0
+
+  name = local.elastic_pool_name
 
   location            = var.location
   resource_group_name = var.resource_group_name
@@ -34,12 +35,12 @@ resource "azurerm_mssql_elasticpool" "elastic_pool" {
   server_name = azurerm_sql_server.server.name
 
   per_database_settings {
-    max_capacity = coalesce(var.database_max_capacity, var.sku.capacity)
-    min_capacity = var.database_min_capacity
+    max_capacity = coalesce(var.elastic_pool_database_max_capacity, var.sku.capacity)
+    min_capacity = var.elastic_pool_database_min_capacity
   }
 
   max_size_gb    = var.elastic_pool_max_size
-  zone_redundant = var.zone_redundant
+  zone_redundant = var.elastic_pool_zone_redundant
 
   sku {
     capacity = local.elastic_pool_sku.capacity
@@ -48,29 +49,7 @@ resource "azurerm_mssql_elasticpool" "elastic_pool" {
     family   = local.elastic_pool_sku.family
   }
 
-  tags = merge(local.default_tags, var.extra_tags, var.server_extra_tags)
-}
-
-resource "azurerm_sql_database" "db" {
-  for_each = var.enable_elasticpool ? toset(var.elasticpool_databases) : []
-
-  name                = each.key
-  location            = var.location
-  resource_group_name = var.resource_group_name
-
-  server_name = azurerm_sql_server.server.name
-  collation   = var.databases_collation
-
-  requested_service_objective_name = "ElasticPool"
-  elastic_pool_name                = azurerm_mssql_elasticpool.elastic_pool[0].name
-
-  threat_detection_policy {
-    email_account_admins = var.enable_advanced_data_security_admin_emails ? "Enabled" : "Disabled"
-    email_addresses      = var.advanced_data_security_additional_emails
-    state                = var.enable_advanced_data_security ? "Enabled" : "Disabled"
-  }
-
-  tags = merge(local.default_tags, var.extra_tags, var.databases_extra_tags)
+  tags = merge(local.default_tags, var.extra_tags, var.elastic_pool_extra_tags)
 }
 
 resource "azurerm_sql_virtual_network_rule" "vnet_rule" {
@@ -79,52 +58,4 @@ resource "azurerm_sql_virtual_network_rule" "vnet_rule" {
   resource_group_name = var.resource_group_name
   server_name         = azurerm_sql_server.server.name
   subnet_id           = each.value.subnet_id
-}
-
-resource "azurerm_mssql_database" "single_database" {
-  for_each  = { for db in local.single_databases_configuration : db.name => db }
-  name      = each.key
-  server_id = azurerm_sql_server.server.id
-
-  sku_name = each.value.sku_name
-
-  collation = each.value.collation
-
-  license_type = lookup(each.value, "license_type", null)
-
-  max_size_gb                 = each.value.max_size_gb
-  zone_redundant              = lookup(each.value, "zone_redundant", false)
-  min_capacity                = lookup(each.value, "min_capacity", null)
-  auto_pause_delay_in_minutes = lookup(each.value, "auto_pause_delay_in_minutes", null)
-
-  dynamic "threat_detection_policy" {
-    for_each = each.value.threat_detection_policy.state == "Enabled" ? ["fake"] : []
-    content {
-      state                = each.value.threat_detection_policy.state
-      email_account_admins = lookup(each.value.threat_detection_policy, "email_account_admins", "Disabled")
-      email_addresses      = lookup(each.value.threat_detection_policy, "email_addresses", "john.doe@azure.com")
-    }
-  }
-
-  short_term_retention_policy {
-    retention_days = each.value.retention_days
-  }
-
-  dynamic "long_term_retention_policy" {
-    for_each = coalesce(
-      lookup(each.value, "weekly_retention", ""),
-      lookup(each.value, "monthly_retention", ""),
-      lookup(each.value, "yearly_retention", ""),
-      lookup(each.value, "week_of_year", ""),
-      "empty"
-    ) == "empty" ? [] : ["fake"]
-    content {
-      weekly_retention  = lookup(each.value, "weekly_retention", null)
-      monthly_retention = lookup(each.value, "monthly_retention", null)
-      yearly_retention  = lookup(each.value, "yearly_retention", null)
-      week_of_year      = lookup(each.value, "week_of_year", null)
-    }
-  }
-
-  tags = merge(local.default_tags, var.extra_tags, lookup(each.value, "database_extra_tags", {}))
 }